main.py can now transform tons of images too
Before my changes made it so we run tons of different variations of
transforms on one image. To find some series of image transforms that
results in a readable image with as few bits as possible.

This commit introduces the first steps to being able to run a choosen
series of transforms on multiple images.

running main.py now still behaves as before however. Not done with the
refactoring yet.