created find_best_steps_one.py
It can be used to find the best series of transforms to do on an image.

main.py now focuses on running such a series on all images.

Until we have OCR integrated we cannot do this all automatically.
We have to run find_best_steps_one.py manually, choose an option that
looks good, add it to main.py and run it there.